+++ fjs/media/json/schema/module.f.ts
@@ -10,7 +10,7 @@ import { array, option, or, record, string } from '../../../types/rtti/module.f.
 import type { Struct, Tuple, Type as RttiType } from '../../../types/rtti/types.ts'
 import { visit } from '../../../types/rtti/common/module.f.mjs'
 import type { Visitor } from '../../../types/rtti/common/types.ts'
-import type { Primitive } from '../../../djs/module.f.ts'
+import type { Primitive } from '../../../djs/types.ts'
 import type { Ts } from '../../../types/rtti/ts/types.ts'
 import type { Phantom } from '../../../types/phantom/types.ts'
 import { unknown as jsonUnknown } from '../rtti/module.f.mjs'