+++ fjs/module.f.ts
@@ -3,7 +3,7 @@
  *
  * @module
  */
-import { compile } from './djs/module.f.ts'
+import { compile } from './djs/module.f.mjs'
 import { main as testMain } from './emergent_testing/module.f.mjs'
 import { commands as casCommands } from './cas/cli/module.f.mjs'
 import { main as ciMain } from './ci/module.f.mjs'